15-simple-agent/pyproject.toml
[project]
name = "15-simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.0.6",
    "python-dotenv>=1.0.1",
]
